+++ NEWS.md
@@ -1,16 +1,18 @@
-# Version 1.3.0.9000
+# Version 1.3.1
 
 ## Added
 * internal functions to serialize and load the session information like connection, process collection, etc.
 
 ## Changed
+* changed OIDC provider and default client selection
+  * automatically selects first OIDC provider withe `default_clients` defined [#142](https://github.com/Open-EO/openeo-r-client/issues/142),
+  * automatically selects a default client by the its id containing "openeo" at first, otherwise the first default client in the list is selected [#146](https://github.com/Open-EO/openeo-r-client/issues/146)
 * argument bounding box tries to extract the EPSG code it EPSGCode was provided as WKT2
 * compute_result now looks for a format in "save_result" if it was not stated in the call
 
 ## Fixes
 * CollectionId argument now checks if a string pattern is available
 * OutputFormat argument checks the "FileFormat" class separately to avoid warnings
-* fixed automated selection of a suitable OIDC provider on login [#142](https://github.com/Open-EO/openeo-r-client/issues/142), where the first OIDC in the list was chosen. Now it checks for the first OIDC provider with `default_clients` defined.
 * fixed an issue where it was not possible to connect to an openEO back-end with a dedicated link (also for reconnect in the connection contract)
 * fixed warning messages shown on `describe_collection()` using an object of class `Collection` from `list_collection()` [#145](https://github.com/Open-EO/openeo-r-client/issues/145)
 * .create_sample_bbox interpretes now correctly a serialized extent object